chore: refactor testdata folder
The idea behind this refactor is to have a better structure for the testdata folder.

It was unclear that each folder in the testdata folder was about activating or deactivating a setting.

For example, the folder `bool-format` was about activating the `bool-format` setting.
But, `err-error` was about activating the setting.

Now, folder names are more explicit about the setting that is being activated or deactivated.